refactor(auth): fix swiftlint
