+++ Tests/InsForgeAuthTests/InsForgeAuthTestSupport.swift
@@ -224,11 +224,11 @@ final class MockURLProtocol: URLProtocol {
     private static var stubs = [Stub]()
     private(set) static var recordedRequests = [URLRequest]()
 
-    override class func canInit(with request: URLRequest) -> Bool {
+    override static func canInit(with request: URLRequest) -> Bool {
         true
     }
 
-    override class func canonicalRequest(for request: URLRequest) -> URLRequest {
+    override static func canonicalRequest(for request: URLRequest) -> URLRequest {
         request
     }
